set default region

+++ src/common.py
@@ -4,7 +4,7 @@ import os
 
 
 s3 = boto3.resource("s3")
-sess = sagemaker.Session(boto3.session.Session())
+sess = sagemaker.Session()
 bucket_name = sess.default_bucket()
 bucket = s3.Bucket(bucket_name)
 region = sess.boto_region_name
@@ -14,8 +14,10 @@ def s3_upload_file(filename, obj_key=None):
         obj_key = filename
     bucket.upload_file(filename, obj_key)
 
-def s3_download_file(obj_key, filename=None):
+def s3_download_file(obj_key, filename=None, if_not_exists=False):
     if filename is None:
         filename = obj_key
+    if if_not_exists and os.path.isfile(filename):
+        return
     with open(filename, 'wb') as f:
         bucket.download_fileobj(obj_key, f)
